chore(ColorRamp): because of macos case issues, we need to rename ColorRamp to ColorRamp_ and back again
