+++ src/ColorRamp_.ts
